--- ui/appui-react/src/test/layout/widget/Widget.test.tsx
@@ -1,40 +1,0 @@
-/*---------------------------------------------------------------------------------------------
- * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
- * See LICENSE.md in the project root for license terms and full copyright notice.
- *--------------------------------------------------------------------------------------------*/
-import { fireEvent, render } from "@testing-library/react";
-import * as React from "react";
-import type { NineZoneDispatch } from "../../../appui-react/layout/base/NineZone.js";
-import { createNineZoneState } from "../../../appui-react/layout/state/NineZoneState.js";
-import { addTab } from "../../../appui-react/layout/state/internal/TabStateHelpers.js";
-import { addFloatingWidget } from "../../../appui-react/layout/state/internal/WidgetStateHelpers.js";
-import {
-  Widget,
-  WidgetIdContext,
-} from "../../../appui-react/layout/widget/Widget.js";
-import { TestNineZoneProvider } from "../Providers.js";
-
-describe("Widget", () => {
-  it("should dispatch FLOATING_WIDGET_BRING_TO_FRONT", () => {
-    const dispatch = vi.fn<NineZoneDispatch>();
-    let state = createNineZoneState();
-    state = addTab(state, "t1");
-    state = addFloatingWidget(state, "w1", ["t1"]);
-    const { container } = render(
-      <TestNineZoneProvider defaultState={state} dispatch={dispatch}>
-        <WidgetIdContext.Provider value="w1">
-          <Widget />
-        </WidgetIdContext.Provider>
-      </TestNineZoneProvider>
-    );
-
-    const widgetElement =
-      container.getElementsByClassName("nz-widget-widget")[0];
-    fireEvent.click(widgetElement);
-
-    expect(dispatch).toHaveBeenCalledWith({
-      type: "FLOATING_WIDGET_BRING_TO_FRONT",
-      id: "w1",
-    });
-  });
-});
